Revert "fix js packages"
This reverts commit 193eb867
